use pyenv

+++ run_once_before_install-01-homebrew.sh
@@ -29,12 +29,16 @@ brew "gawk"
 brew "pyenv"
 EOF
 
-PYTHON_VERSION=$(pyenv install --list | grep --extended-regexp "^\s*[0-9][0-9.]*[0-9]\s*$" | tail -1)
-pyenv install $PYTHON_VERSION 
-pyenv global $PYTHON_VERSION
+if [ -f "$HOME/.pyenv" ]; then
+    PYTHON_VERSION=$(pyenv install --list | grep --extended-regexp "^\s*[0-9][0-9.]*[0-9]\s*$" | tail -1)
+    pyenv install $PYTHON_VERSION 
+    pyenv global $PYTHON_VERSION
+    pip install --upgrade pip
+    pip install --user pipenv
+    echo "[CHEZMOI] 🐍 Installed Python $PYTHON_VERSION with pyenv"
+fi
 
-pip install --upgrade pip
-pip install --user pipenv
+
 
 if [ -f "$HOME/bin/chezmoi" ]; then
     echo "[CHEZMOI] 🍺 Delete chezmoi bin (which is now managed by homebrew)"
